Authorization with JWT Authorization with Google

+++ handlers/tasks.py
@@ -1,10 +1,11 @@
 from typing import Annotated
-from fastapi import APIRouter, status, Depends
-from scheme.task import TaskSchema
+from fastapi import APIRouter, status, Depends, HTTPException
+from scheme.task import TaskSchema, TaskCreateSchema
+from models import UserProfile
 from service import TaskService
 from repository import TasksRepository
 
-from dependency import get_task_repository, get_task_service
+from dependency import get_task_repository, get_task_service, get_current_user
 
 router = APIRouter(prefix='/task', tags=['task'])
 
@@ -14,18 +15,23 @@ router = APIRouter(prefix='/task', tags=['task'])
     response_model=list[TaskSchema]
 )
 async def get_tasks(
-        task_service: Annotated[TaskService, Depends(get_task_service)]
+    task_service: Annotated[TaskService, Depends(get_task_service)],
+    current_user: Annotated[UserProfile, Depends(get_current_user)]
 ):
-    return task_service.get_tasks()
+    return task_service.get_tasks(user_id=current_user.id)
 
 
 @router.post(
     '/',
     response_model=TaskSchema
 )
-async def create_task(task: TaskSchema,
-                      task_repository: Annotated[TasksRepository, Depends(get_task_repository)]):
-    task_id = task_repository.create_task(task)
+async def create_task(
+    task: TaskCreateSchema,
+    task_repository: Annotated[TasksRepository, Depends(get_task_repository)],
+    current_user: Annotated[UserProfile, Depends(get_current_user)]
+):
+    task.user_id = current_user.id  # Привязываем задачу к пользователю
+    task_id = task_repository.create_task(task, current_user.id)
     task.id = task_id
     return task
 
@@ -37,8 +43,16 @@ async def create_task(task: TaskSchema,
 async def update_task(
         task_id: int,
         name: str,
-        task_repository: Annotated[TasksRepository, Depends(get_task_repository)]
+        task_repository: Annotated[TasksRepository, Depends(get_task_repository)],
+        current_user: Annotated[UserProfile, Depends(get_current_user)]  # Проверяем, что пользователь авторизован
 ):
+    task = task_repository.get_task(task_id)
+
+    if not task:
+        raise HTTPException(status_code=404, detail="Task not found")
+
+    if task.user_id != current_user.id:
+        raise HTTPException(status_code=403, detail="You can only update your own tasks")
 
     return task_repository.update_task_name(task_id, name)
 
@@ -49,7 +63,17 @@ async def update_task(
 )
 async def delete_task(
         task_id: int,
-        task_repository: Annotated[TasksRepository, Depends(get_task_repository)]
+        task_repository: Annotated[TasksRepository, Depends(get_task_repository)],
+        current_user: Annotated[UserProfile, Depends(get_current_user)]  # Проверяем, что пользователь авторизован
 ):
+    task = task_repository.get_task(task_id)
+
+    if not task:
+        raise HTTPException(status_code=404, detail="Task not found")
+
+    if task.user_id != current_user.id:
+        raise HTTPException(status_code=403, detail="You can only delete your own tasks")
+
     task_repository.delete_task(task_id)
-    return {'message': 'task deleted successfully'}
+    return {'message': 'Task deleted successfully'}
+
